requirements.txt
absl-py==0.7.1
aiohttp==3.5.4
APScheduler==3.6.1
astor==0.8.0
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
chart-studio==1.0.0
Click==7.0
colorama==0.4.1
cycler==0.10.0
docopt==0.6.2
filelock==3.0.12
Flask==1.1.1
flatbuffers==1.11
funcsigs==1.0.2
gast==0.2.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
hdfs==2.5.8
idna==2.8
idna-ssl==1.1.0
importlib-metadata==0.19
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mock==3.0.5
more-itertools==7.2.0
multidict==4.5.2
mysql-connector-python==8.0.17
numpy==1.16.5
packaging==19.1
pandas==0.24.2
pathlib2==2.3.4
pika==1.1.0
plotly==4.1.1
pluggy==0.12.0
protobuf==3.9.1
psutil==5.6.3
py==1.8.0
pymongo==3.9.0
pyparsing==2.4.2
pytest==5.1.1
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
ray==0.7.1
redis==3.3.8
requests==2.22.0
retrying==1.3.3
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
six==1.12.0
sklearn==0.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
typing==3.7.4.1
typing-extensions==3.7.4
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.5
wrapt==1.11.2
yarl==1.3.0
zipp==0.6.0
